third_party/blink/web_tests/fullscreen/full-screen-with-flex-item-expected.txt

Enter Fullscreen
EVENT(webkitfullscreenchange)
EVENT(webkitfullscreenchange)
EXPECTED (600 == '600') OK
END OF TEST
